Fix macOS build: Roaring64Map ambiguity, glibc-only headers, endianness
On Darwin size_t is unsigned long while uint64_t is unsigned long long,
making Roaring64Map add/contains/addRange calls ambiguous; cast call
sites explicitly (same treatment as the fix-macos-build branch).
Replace bits/stdc++.h and bits/types/struct_iovec.h with standard
headers, detect endianness on __APPLE__, guard EOF for nlohmann after
the ANTLR runtime #undefs it, align new_iovec definition with its
declaration, and rename a template parameter that shadows std::size.

Re-pin nested submodules to commits that exist upstream and work with
RTTI enabled: antlr4 back to the stock 4.11.1 release (the no-rtti
patched runtime replaces dynamic_cast with C-style casts, breaking
getRuleContexts type filtering and any visitor-based grammar), json
v3.11.3 (recorded pin unreachable), magic_enum v0.9.7 (the code uses
the nested include layout the old flat pin predates).

+++ src/yaucl/data/new_iovec.cpp
@@ -4,7 +4,7 @@
 
 #include "yaucl/data/new_iovec.h"
 
-new_iovec::new_iovec(void *memory, size_t size) : iov_base(memory), iov_len(size) {}
+new_iovec::new_iovec(void *memory, uint_fast64_t size) : iov_base(memory), iov_len(size) {}
 
 std::istringstream new_iovec::stream() {
     std::string elem2{(char*)iov_base,iov_len};
